+++ src/test/java/de/samply/converter/ConverterManagerTest.java
@@ -4,6 +4,7 @@ import de.samply.csv.ContainersToCsvConverter;
 import de.samply.fhir.BundleToContainersConverter;
 import de.samply.template.ConverterTemplate;
 import de.samply.template.ConverterTemplateManager;
+import java.util.Set;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import reactor.core.publisher.Flux;
@@ -23,10 +24,17 @@ class ConverterManagerTest {
     ConverterManager converterManager = new ConverterManager(bundleToContainersConverter, containersToCsvConverter, CONVERTER_APPLICATION_CONTEXT_PATH);
     Converter converter = converterManager.getBestMatchConverter(Format.FHIR_QUERY, Format.CSV, sourceId);
     ConverterTemplateManager converterTemplateManager = new ConverterTemplateManager("./templates");
+
+    Set<String> sourceIds = converterManager.getSourceIds();
+    //TODO
+    Set<String> converterTemplateIds = converterTemplateManager.getConverterTemplateIds();
+    //TODO
+
     ConverterTemplate converterTemplate = converterTemplateManager.getConverterTemplate("test-template1");
     Flux flux = converter.convert(Flux.just("Patient"), converterTemplate);
     flux.blockLast();
     //TODO
+
   }
 
 }